Happy place with components...
I've tested a few things with components, trying to autoload them,
trying to find a wait to make them a bit more automated, the conclusion
I have come to on this is that its the wrong approach, it opens a
plethora of issues surrounding them.

The very definition of the component is a long lasting over arching
struct that can run independently of the engine itself while providing
functionality to the engine in the form of events. By this definition I
feel lie the automated approach is just wrong, I've simplified them a
little bit as we don't need it to provide a run method, the component
can do this independently. For the times we need something closely
coupled to an engine object or client, we should be using Traits.